requirements.txt
PyNiteFEA>=0.0.94
numpy>=1.21.0
matplotlib>=3.5.0
